[refactor]: Abstract functions into separate modules to improve code quality and scalability
Functions are now categorized into two;
helper functions in helpers.js and
authentication functions in authenticate.js
