+++ helpers.js
@@ -1,0 +1,54 @@
+import NodeRSA from 'node-rsa';
+import jwt from 'jsonwebtoken';
+import bodyParser from 'body-parser';
+
+/**
+ * Get PEM encoded RSA private key
+ * function generateToken() uses key for token creation
+ * @function generateRSAPrivateKey
+ * @returns {String} PEM-encoded RSA private key
+ */
+const generateRSAPrivateKey = () => {
+  const key = new NodeRSA({ b: 1024 }, 'pkcs1-private-pem');
+  const exported = key.exportKey('pkcs1-private-pem');
+  return exported;
+};
+
+/**
+ * Generate json web token
+ * @function generateToken
+ * @param {Object} user - User object returned from firebase authentication
+ * @param {String} rsaKey - RSA private key generated by
+ *  generateRSAPrivateKey()
+ * @returns {String} - A json web token
+ */
+const generateToken = (user, rsaKey) => {
+  const privateClaim = { uid: user.uid };
+  const options = {
+    algorithm: 'RS256',
+    issuer: 'expressFireAuth',
+    subject: user.email,
+    expiresIn: '30d'
+  };
+  const token = jwt.sign(privateClaim, rsaKey, options);
+  return token;
+};
+
+/**
+ * Parse the body of requests and populate req.body with payload
+ * See {@link https://expressjs.com/en/4x/api.html#req.body ExpressJS}
+ * @function parseRequestBody
+ * @returns {Array} - An array of two body-parser express middlewares
+ */
+const parseRequestBody = () => {
+  return [
+    bodyParser.json(),
+    bodyParser.urlencoded({ extended: 'true' })
+  ];
+};
+
+export {
+  generateToken,
+  parseRequestBody,
+  generateRSAPrivateKey,
+};
